+++ pkg/drift_client/drift_client.py
@@ -96,18 +96,18 @@ class DriftClient:
             >>> # => {"topic-1": ['topic-1/1644750600291.dp',
             >>> #                  'topic-1/1644750601291.dp', ...] ... }
         """
-        to_request = self.__influx_client.query_data(topics, timeframe[0], timeframe[1])
-
         data = {}
-
-        if not to_request:
-            return data
-
         for topic in topics:
-            data[topic] = []
+            influxdb_values = self.__influx_client.query_data(
+                topic, timeframe[0], timeframe[1], field="status"
+            )
 
-        for file in to_request:
-            data[file.split("/")[0]].append(file)
+            if not influxdb_values:
+                break
+
+            data[topic] = []
+            for timestamp, _ in influxdb_values:
+                data[topic].append(f"{topic}/{int(timestamp * 1000)}.dp")
 
         return data