+++ tests/drift_client_test.py
@@ -13,6 +13,13 @@ def _mock_influx_class(mocker):
     return mocker.patch("drift_client.drift_client.InfluxDBClient")
 
 
+@pytest.fixture(name="influxdb_client")
+def _mock_influx_client(mocker, influxdb_klass):
+    client = mocker.Mock()
+    influxdb_klass.return_value = client
+    return client
+
+
 def test__default_initialization(minio_klass, influxdb_klass):
     """should initialize clients with default settings"""
     _ = DriftClient("host_name", "password")
@@ -20,4 +27,19 @@ def test__default_initialization(minio_klass, influxdb_klass):
     minio_klass.assert_called_with("http://host_name:9000", "panda", "password", False)
     influxdb_klass.assert_called_with(
         "http://host_name:8086", "panda", "password", False
+    )
+
+
+@pytest.mark.usefixtures("minio_klass")
+def test__timestamp_from_influxdb(influxdb_client):
+    """should get timestamp and values for records
+    from influxdb and make paths in minio"""
+    client = DriftClient("host_name", "password")
+    influxdb_client.query_data.return_value = [(10000.0, 0), (10010.0, 512)]
+
+    data = client.get_list(["topic"], ["2022-01-01 00:00:00", "2022-01-01 00:00:00"])
+    assert data == {"topic": ["topic/10000000.dp", "topic/10010000.dp"]}
+
+    influxdb_client.query_data.assert_called_with(
+        "topic", "2022-01-01 00:00:00", "2022-01-01 00:00:00", field="status"
     )